Semestral work: Change whole implementation to use integers in vector2D

+++ Aurorion/Sources/CCamera.cpp
@@ -44,7 +44,7 @@ void CCamera::Update(float deltaTime) {
     if (m_ViewPort.x > TheGame::Instance().GetMapWidth() - TheGame::Instance().GetWindowWidth())
         m_ViewPort.x = TheGame::Instance().GetMapWidth() - TheGame::Instance().GetWindowWidth();
 
-    m_Position = {static_cast<float>(m_ViewPort.x), static_cast<float>(m_ViewPort.y)};
+    m_Position = {m_ViewPort.x, m_ViewPort.y};
 //    std::cout << "TARGET : " << m_Target->GetX() << " " << m_Target->GetY() << std::endl;
 //    std::cout << "VIEW : " << m_ViewPort.x << " " << m_ViewPort.y << std::endl;
 //    std::cout << "POS: " << m_Position.GetX() << m_Position.GetY() << std::endl;
